fixed oAuth
updated ASWebAuthenticationSession implementation for iOS 13

+++ proto-application/proto-application/helper files/news.swift
@@ -27,7 +27,7 @@ struct NewsElement: Codable {
     let title: String?
     let featuredImageURL: String?
     let content: String?
-    let publishedAt: Int?
+    let publishedAt: Double?
     
     enum CodingKeys: String, CodingKey {
         case id, title
